Crate junction

Library for working with NTFS junctions.

Junction Points are a little known NTFS v5+ feature roughly equivalent to Unix directory symbolic links.

They are supported in Windows 2000 and onwards, where a directory serves as a symbolic link to another directory on the computer. For example, if the directory D:\SYMLINK specified C:\WINNT\SYSTEM32 as its target, then an application accessing D:\SYMLINK\DRIVERS would in reality be accessing C:\WINNT\SYSTEM32\DRIVERS.

Functions

create

Creates a junction point from the specified directory to the specified target directory.

delete

Deletes a junction reparse point from the specified file or directory.

exists

Determines whether the specified path exists and refers to a junction point.

get_target

Gets the target of the specified junction point.